remove round trip through OffTopology
mainly I'm not sure how to do this because an openmm::Topology is an opaque
wrapper around a C++ type from openmm and an OffTopology is going to be a
wrapper around an opaque Python type. I don't think the Python and C++ versions
can talk directly to each other like that, so I would have to do two conversions
to real Rust structs and then implement From. it appears this round trip is just
to detect virtual sites anyway